Python codes/requirements.txt
gradio
torch
diffusers
transformers
huggingface_hub
Pillow
numpy
opencv-python
